doc/pyproject.toml
[project]
name = "CSE-documentation"
version = "0.1.0"
description = "Source code for CSE documentation website"
authors = [{ name = "The CSE Authors" }]
license = "BSD-3"
readme = "README.md"
keywords = ["cse", "energy-modeling", "simulation", "documentation"]
requires-python = ">=3.12"
dependencies = [
    "mkdocs-material",
    "mkdocs-autorefs>=1.3.1",
    "mkdocs-macros-plugin>=1.3.7",
    "mkdocs-awesome-nav>=3.1.1",
    "pyyaml>=6.0.2",
    "python-slugify>=8.0.4",
]


[tool.uv]
dev-dependencies = [
    "doit",
    "ruff>=0.9.9",
]

[tool.mypy]
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
